Separación de código y ordenación
